pyproject.toml
[project]
name = "flow-meeting-teams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=1.7.2",
    "langchain-openai>=0.3.23",
    "openai>=1.83.0",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.10.1",
    "python-docx>=1.2.0",
    "streamlit>=1.52.2",
]
